tests/wpt/tests/css/css-pseudo/parsing/the-select-arrow-pseudo-element.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS ::select-arrow Pseudo-Element Test</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/pull/10986">
<meta name="assert" content="This test checks the validity of the ::select-arrow pseudo element selector." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_valid_selector("::select-arrow");
test_valid_selector("*::select-arrow", "::select-arrow");
test_valid_selector("foo.bar[baz]::select-arrow");
test_invalid_selector("::select-arrow *");

// Combinations
test_invalid_selector("::select-arrow::select-arrow");

test_invalid_selector("::before::select-arrow");
test_invalid_selector("::after::select-arrow");
test_invalid_selector("::marker::select-arrow");
test_invalid_selector("::placeholder::select-arrow");

test_invalid_selector("::select-arrow::before");
test_invalid_selector("::select-arrow::after");
test_valid_selector("::select-arrow::marker");
test_invalid_selector("::select-arrow::placeholder");

test_invalid_selector("::slotted(*)::select-arrow::slotted(*)");
test_valid_selector("::slotted(*)::select-arrow");

test_valid_selector("::part(foo)::select-arrow");
</script>
